requirements_multilang.txt
# ==================== CORE DEPENDENCIES ====================
langchain>=0.2.0
langchain-openai>=0.1.0
langchain-community>=0.2.0
langchain-chroma>=0.1.0

# ==================== AI/ML ====================
chromadb>=0.4.0
dashscope>=1.14.0

# ==================== MULTI-LANGUAGE SUPPORT ====================
# Language detection
langdetect>=1.0.9

# Translation
googletrans==4.0.0rc1

# Unicode handling
unicodedata2>=15.0.0

# Language-specific fonts and rendering
fonttools>=4.42.0

# ==================== DOCUMENT PROCESSING ====================
python-docx>=1.1.0
pandas>=2.0.0
openpyxl>=3.1.0
tabulate>=0.9.0
unstructured>=0.10.0
unstructured[docx]
unstructured[pdf]

# PDF handling for multi-language
pdfplumber>=0.10.0
reportlab>=4.0.0  # For PDF generation with Unicode support

# ==================== WEB INTERFACE ====================
streamlit>=1.28.0
plotly>=5.14.0

# ==================== UTILITIES ====================
python-dotenv>=1.0.0
requests>=2.31.0
tqdm>=4.65.0
colorlog>=6.7.0
watchdog>=3.0.0

# ==================== DATA VISUALIZATION ====================
matplotlib>=3.7.0
seaborn>=0.12.0

# ==================== CACHING ====================
redis>=4.5.0  # Optional for distributed caching